[cmake] refactor: cpmfile, deps prefetch, force system and more
This is a (currently) very WIP PR that basically completely refactors
the CPM system for the third time. Dependencies are now managed in a
singular json file, where each can be properly prefetched at-will via a
script I'll add eventually. For the time being, this is super incomplete
and I have to go to class. Once I'm back I'll keep working on it.

Signed-off-by: crueter <crueter@eden-emu.dev>